chore: 예시 코드의 객체 분리 (#21)
