+++ src/pages/home/Home.tsx
@@ -1,40 +1,21 @@
 import React from "react";
 
-import { Button, Carousel, Input, Modal, Profile, Toggle } from "components";
+import {
+  Button,
+  Carousel,
+  Input,
+  Modal,
+  Profile,
+  Tab,
+  Toggle,
+} from "components";
 import { useModal, useToast } from "hooks";
 import Img from "assets/images/img_01.jpg"; //TODO:임시 저장 이미지
+import { toastContext, carouselData, info, keywords } from "./HomeDummy";
 
 const Home = () => {
   const { addToast } = useToast();
   const { handleOpenModal } = useModal();
-
-  const toastContext = {
-    content: "토스트 테스트중",
-  };
-
-  const carouselData = [
-    {
-      id: 1,
-      title: "첫번째 타이틀",
-      content: "첫번째 요소",
-      img: "vite.svg",
-    },
-    {
-      id: 2,
-      title: "두번째 타이틀",
-      content: "두번째 요소",
-      img: "vite.svg",
-    },
-    {
-      id: 3,
-      title: "세번째 타이틀",
-      content: "세번째 요소",
-      img: "vite.svg",
-    },
-  ];
-
-  const info = { location: "서울", company: "회사", job: "직업" };
-  const keywords = ["키", "특징", "특징"];
 
   return (
     <>
@@ -60,6 +41,7 @@ const Home = () => {
         info={info}
         keywords={keywords}
       />
+      <Tab />
     </>
   );
 };
